ai-rag-chatbot/requirements.txt
python-dotenv
ibm-watsonx-ai
langchain-ibm
langchain-text-splitters
langchain-community
langchain-classic
gradio
pypdf
chromadb
